+++ HR-Client/src/app/components/document-detail/document-detail.component.html
