+++ HR-Client/src/app/services/visa-status.service.ts
@@ -13,7 +13,7 @@ export class HrVisaStatusService {
 
   constructor(private http: HttpClient) {}
   private getHttpOptions() {
-    const token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NWUyNWU3ZWE4MDlhNzI4NjZiNWQzNzMiLCJpYXQiOjE3MDk0MzUzMzYsImV4cCI6MTcwOTQzODkzNn0.LU8E6aiK4KmqcOQt51UwQaUaaDKIKEWnou3K5T0e9ds";
+    const token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NWUyNWU3ZWE4MDlhNzI4NjZiNWQzNzMiLCJpYXQiOjE3MDk0NDgwMjgsImV4cCI6MTcwOTQ1MTYyOH0.rH5ZZqPi32u6Te7d2q2vMUbPfbRsw5vN5KeuEw4Tb1M";
     return {
       headers: new HttpHeaders({
         'Authorization': `Bearer ${token}`
